+++ src/main/java/com/parse/gwt/client/types/GeoPoint.java
@@ -1,0 +1,9 @@
+package com.parse.gwt.client.types;
+
+/**
+ * Created by Kerby on 2/23/2017.
+ */
+public class GeoPoint {
+    public Double longitude;
+    public Double latitude;
+}