+++ src/main/java/com/promis/rtti/annotations/NoAssign.java
@@ -1,0 +1,43 @@
+/*
+ * The contents of this file are subject to the Mozilla Public License
+ * Version 1.1 (the "License"); you may not use this file except in
+ * compliance with the License. You may obtain a copy of the License at
+ * http://www.mozilla.org/MPL/
+ *
+ * Software distributed under the License is distributed on an "AS IS"
+ * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See the
+ * License for the specific language governing rights and limitations
+ * under the License.
+ *
+ * The Original Code is GWT RTTI library.
+ *
+ * The Initial Developer of the Original Code is
+ * Jan "SHadoW" Rames <ramejan@gmail.com>.
+ * Portions created by the Initial Developer are Copyright (C) 2011
+ * the Initial Developer. All Rights Reserved.
+ */
+package com.promis.rtti.annotations;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import com.promis.rtti.client.UnableToAssignException;
+
+/**
+ * Identifies a class for which no <code>assign</code>
+ * code will be generated. In which case <code>assign</code>
+ * always returns <code>false</code> (throws 
+ * {@link UnableToAssignException}). 
+ * @author SHadoW
+ *
+ */
+@Documented  
+@Retention(RetentionPolicy.RUNTIME)  
+@Target({ElementType.TYPE, ElementType.FIELD})  
+public @interface NoAssign
+{
+
+}